IViewLocator implementations now access view providers in reverse order of registration.
