+++ CoreMVVM.Demo/ViewProvider.cs
@@ -1,10 +1,12 @@
-﻿using CoreMVVM.Demo.Views;
-using System;
+﻿using System;
+using System.Collections.Generic;
 
 namespace CoreMVVM.Demo
 {
     public sealed class ViewProvider : IViewProvider
     {
+        private readonly Dictionary<Type, Type> _registeredViews = new Dictionary<Type, Type>();
+
         public Type FindView<TViewModel>()
         {
             return FindView(typeof(TViewModel));
@@ -12,14 +14,18 @@ namespace CoreMVVM.Demo
 
         public Type FindView(Type viewModel)
         {
-            if (viewModel == typeof(DialogWindowModel))
-                return typeof(DialogWindow);
-
             if (viewModel is null)
                 throw new ArgumentNullException(nameof(viewModel));
 
-            string viewTypeName = viewModel.Namespace.Replace("ViewModel", "View").Replace("WindowModel", "Window");
-            return viewModel.Assembly.GetType(viewTypeName);
+            if (_registeredViews.TryGetValue(viewModel, out Type viewType))
+                return viewType;
+
+            return null;
+        }
+
+        public void RegisterView<TViewModel, TView>()
+        {
+            _registeredViews[typeof(TViewModel)] = typeof(TView);
         }
     }
 }
