+++ CoreMVVM/Implementations/ViewLocator.cs
@@ -1,6 +1,7 @@
 ﻿using CoreMVVM.IOC;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Reflection;
 
 namespace CoreMVVM.Implementations
@@ -42,14 +43,7 @@ namespace CoreMVVM.Implementations
         {
             _logger.Debug($"View for view model '{typeof(TViewModel)} requested.");
 
-            Type viewType = null;
-            foreach (var viewProvider in _viewProviders)
-            {
-                viewType = viewProvider.FindView(typeof(TViewModel));
-                if (viewType != null)
-                    break;
-            }
-
+            Type viewType = LocateViewType(provider => provider.FindView<TViewModel>());
             if (viewType is null)
             {
                 _logger.Error($"Failed to find view for view model of type '{typeof(TViewModel)}'.");
@@ -68,14 +62,8 @@ namespace CoreMVVM.Implementations
             _logger.Debug($"View for view model '{viewModel.GetType()} requested.");
 
             Type viewModelType = viewModel.GetType();
-            Type viewType = null;
-            foreach (var viewProvider in _viewProviders)
-            {
-                viewType = viewProvider.FindView(viewModelType);
-                if (viewType != null)
-                    break;
-            }
 
+            Type viewType = LocateViewType(provider => provider.FindView(viewModelType));
             if (viewType is null)
             {
                 _logger.Error($"Failed to find view for view model of type '{viewModel.GetType()}'.");
@@ -99,6 +87,18 @@ namespace CoreMVVM.Implementations
         #endregion IViewLocator
 
         #region Helpers
+
+        private Type LocateViewType(Func<IViewProvider, Type> locator)
+        {
+            foreach (var viewProvider in Enumerable.Reverse(_viewProviders))
+            {
+                Type viewType = locator(viewProvider);
+                if (viewType != null)
+                    return viewType;
+            }
+
+            return null;
+        }
 
         private object CreateView(Type viewType, object viewModel)
         {
@@ -134,22 +134,5 @@ namespace CoreMVVM.Implementations
         }
 
         #endregion Helpers
-    }
-
-    /// <summary>
-    /// A basic implementation of the view provider.
-    /// </summary>
-    public sealed class DefaultViewProvider : IViewProvider
-    {
-        public Type FindView<TViewModel>()
-        {
-            return FindView(typeof(TViewModel));
-        }
-
-        public Type FindView(Type viewModel)
-        {
-            string viewTypeName = viewModel.FullName.Replace("ViewModel", "View").Replace("WindowModel", "Window");
-            return viewModel.Assembly.GetType(viewTypeName);
-        }
     }
 }
